+++ udev.py
@@ -70,6 +70,16 @@
     provide access to udev attributes of the device (like its path in
     ``sysfs``).
 
+    You can not only list existing devices, you can also monitor the device
+    list for changes using the :class:`Monitor` class:
+
+    >>> monitor = Monitor.from_netlink(context)
+    >>> monitor.filter_by(subsystem='input')
+    >>> for action, device in monitor:
+    ...     if action == 'add':
+    ...         print device, 'added'
+
+
     Remarks
     -------
 
@@ -80,8 +90,10 @@
 """
 
 import sys
+import select
 from itertools import count
 from collections import Mapping
+from contextlib import closing
 
 import _udev
 
@@ -436,3 +448,202 @@ class Device(Mapping):
             return self.device_path != other.device_path
         else:
             return self.device_path != other
+
+
+class Monitor(object):
+    """
+    Monitor udev events:
+
+    >>> context = Context()
+    >>> monitor = Monitor.from_netlink(context)
+    >>> monitor.filter_by(subsystem='input')
+    >>> for action, device in monitor:
+    ...     print action, device
+    ...
+
+    A :class:`Monitor` objects connects to the udev daemon and listens for
+    changes to the device list.  A monitor is created by connecting to the
+    kernel daemon through netlink (see :meth:`from_netlink`).
+    Alternatively, connections to arbitrary daemons can be made using
+    :meth:`from_socket`, which is however only seldom of use.
+
+    Once the monitor is created, you can add a filter using
+    :meth:`filter_by` to drop incoming events in subsystems, which are not
+    of interest to the application.
+
+    If the monitor is eventually set up, you can either iterate over the
+    :class:`Monitor` object.  In this case, the monitor implicitly starts
+    listening, and polls for incoming events.  Such events are then yielded
+    to the caller.  Iteration is a blocking operation and does not integrate
+    into external event loops.  If such integration is required, you can
+    explicitly enable the monitor (see :meth:`enable_receiving`), and then
+    retrieve a file description using :meth:`fileno`.  This file descriptor
+    can then be passed to classes like ``QSocketNotifier`` from Qt4.
+    """
+
+    def __init__(self, context, monitor_p):
+        self.context = context
+        self._monitor = monitor_p
+
+    def __del__(self):
+        libudev.udev_monitor_unref(self._monitor)
+
+    @classmethod
+    def from_netlink(cls, context, source='udev'):
+        """
+        Create a monitor by connecting to the kernel daemon through netlink.
+
+        ``context`` is the :class:`Context` to use.  ``source`` is the event
+        source.  As of now, two sources are available:
+
+        - ``'udev'`` (the default):  Events emitted after udev as registered
+          and configured the device.  This is the absolutely recommended
+          source for applications.
+        - ``'kernel'``:  Events emitted directly after the kernel has seen
+          the device.  The device has not yet been configured by udev and
+          might not be usable at all.  **Never** use this, unless you know
+          what you are doing.
+
+        Return a new :class:`Monitor` object, which is connected to the
+        given source.  Raise :exc:`~exceptions.ValueError`, if an invalid
+        source has been specified.  Raise
+        :exc:`~exceptions.EnvironmentError`, if the creation of the monitor
+        failed.
+        """
+        if source not in ('kernel', 'udev'):
+            raise ValueError('Invalid source: {0!r}. Must be one of "udev" '
+                             'or "kernel"'.format(source))
+        source = _assert_bytes(source)
+        monitor = libudev.udev_monitor_new_from_netlink(
+            context._context, source)
+        if not monitor:
+            raise EnvironmentError('Could not create udev monitor')
+        return cls(context, monitor)
+
+    @classmethod
+    def from_socket(cls, context, socket_path):
+        """
+        Connect to an arbitrary udev daemon using the given ``socket_path``.
+
+        ``context`` is the :class:`Context` to use. ``socket_path`` is
+        string, pointing to an existing socket.  If the path starts with a
+        @, use an abstract namespace socket.  If ``socket_path`` does not
+        exist, fall back to an abstract namespace socket.
+
+        The caller is responsible for permissions and cleanup of the socket
+        file.
+
+        Return a new :class:`Monitor` object, which is connected to the
+        given socket.  Raise :exc:`~exceptions.EnvironmentError`, if the
+        creation of the monitor failed.
+        """
+        socket_path = _assert_bytes(socket_path)
+        monitor = libudev.udev_monitor_new_from_socket(
+            context._context, source)
+        if not monitor:
+            raise EnvironmentError('Could not create udev monitor')
+        return cls(context, monitor)
+
+    def fileno(self):
+        """
+        Return the file description associated with this monitor as integer.
+
+        This is really a real file descriptor ;), which can be watched and
+        :func:`select.select`\ ed.
+        """
+        return libudev.udev_monitor_get_fd(self._monitor)
+
+    def filter_by(self, subsystem=None, device_type=None):
+        """
+        Filter incoming events.
+
+        ``subsystem`` can either be ``None`` to not filter against any
+        specific subsystem, or a byte or unicode string with the name of a
+        subsystem (e.g. ``'input'``).  In the latter case, only events
+        originating from the given subsystem pass the filter and are given
+        to the caller.
+
+        ``device_type`` can either be ``None`` to not filter against any
+        specific type of device, or a byte or unicode string specifying the
+        device type.  In the latter case, only events originating from
+        devices with the given type pass the filter and are given to the
+        caller.
+
+        If both arguments are specified, only events from the given
+        subsystem *and* with the given device type are accepted.
+
+        This method must be called *before* :meth:`enable_receiving`.
+
+        Raise :exc:`~exceptions.EnvironmentError`, if the filter could not
+        be installed.
+        """
+        if subsystem:
+            subsystem = _assert_bytes(subsystem)
+        if device_type:
+            device_type = _assert_bytes(device_type)
+        error = libudev.udev_monitor_filter_add_match_subsystem_devtype(
+            self._monitor, subsystem, device_type)
+        if error:
+            raise EnvironmentError('Could not install filter on monitor')
+
+    def enable_receiving(self):
+        """
+        Switch the monitor into listing mode.
+
+        Connect to the event source and receive incoming events.  Only after
+        calling this method, the monitor listens for incoming events.
+
+        .. note::
+
+           This method is implicitly called by :meth:`__iter__`.  You don't
+           need to call it explicitly, if you are iterating over the
+           monitor.
+        """
+        error = libudev.udev_monitor_enable_receiving(self._monitor)
+        if error:
+            raise EnvironmentError('Could not put monitor into '
+                                   'receiving mode')
+
+    start = enable_receiving
+
+    def receive_device(self):
+        """
+        Receive a single device from the monitor.
+
+        The caller must make sure, that there are events available in the
+        event queue.  The call may block, until a device is available.
+
+        If a device was available, return ``(action, device)``.  ``action``
+        is a string describing the action (e.g. ``'add'``, ``'remove'``).
+        ``device`` is :class:`Device` object describing the device.  If no
+        device was available, return ``None``.
+        """
+        device_p = libudev.udev_monitor_receive_device(self._monitor)
+        if not device_p:
+            return None
+        action = libudev.udev_device_get_action(device_p).decode(
+            sys.getfilesystemencoding())
+        return action, Device(self.context, device_p)
+
+    def __iter__(self):
+        """
+        Wait for incoming events and receive them upon arrival.
+
+        This methods implicitly calls :meth:`enable_receiving`, and starts
+        polling the :meth:`fileno` of this monitor.  If a event comes in, it
+        receives the corresponding device and yields it to the caller.
+
+        The returned iterator is endless, and continues receiving devices
+        without ever stopping.
+
+        Yields ``(action, device)`` (see :meth:`receive_device` for a
+        description).
+        """
+        self.enable_receiving()
+        with closing(select.epoll()) as notifier:
+            notifier.register(self, select.EPOLLIN)
+            while True:
+                events = notifier.poll()
+                if events:
+                    yield self.receive_device()
+